deleted obsolete example
