+++ tests/test_examples.py
@@ -74,24 +74,10 @@ def test_basic_household_with_all_resultfiles():
 #     hisim_main.main(path, func, mysimpar)
 @utils.measure_execution_time
 def test_modular_household_configurations( ):
-    # path = "../examples/modular_household.py"
-    # func = "modular_household_explicit"
-    # mysimpar = SimulationParameters.one_day_only( year = 2019, seconds_per_timestep = 60 ) 
-    # # for pv_included in [ True, False ]:
-    # #     for smart_devices_included in [ True, False ]:
-    # #         for boiler_included in [ 'electricity', 'hydrogen', None ]:
-    # #             for heating_device_included in [ 'heat_pump', 'oil_heater', 'district_heating' ]:
-    # predictive = random.choice( [ True, False ] )
-    # pv_included = random.choice( [ True, False ] )
-    # smart_devices_included = random.choice( [ True, False ] )
-    # boiler_included = random.choice( [ 'electricity', 'hydrogen', None ] )
-    # heating_device_included = random.choice( [ 'heat_pump', 'oil_heater', 'district_heating' ] )
-    # mysimpar.reset_system_config( predictive = predictive,
-    #                               pv_included = pv_included, 
-    #                               smart_devices_included = smart_devices_included, 
-    #                               boiler_included = boiler_included, 
-    #                               heating_device_included = heating_device_included )
-    # hisim_main.main( path, func, mysimpar )
+    path = "../examples/modular_household.py"
+    func = "modular_household_explicit"
+    mysimpar = SimulationParameters.one_day_only( year = 2019, seconds_per_timestep = 60 )
+    hisim_main.main( path, func, mysimpar )
     pass
 @utils.measure_execution_time
 def test_first_example():
